checking req_id matches in responses where available

+++ src/main/java/com/cb/model/kraken/ws/response/KrakenError.java
@@ -11,7 +11,7 @@ public class KrakenError {
 
     private String error;
     private String method;
-    private Long req_id;
+    private Integer req_id;
     private boolean success;
     private Instant time_in;
     private Instant time_out;
